pyproject.toml
[tool.poetry]
name = "zenhr"
version = "0.1.0"
description = "A CRUD app for ZipHR Airlines"
authors = ["Vlad Tabakov <vl.tab.kov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
django-environ = "^0.9.0"
psycopg2 = "^2.9.3"
drf-yasg = "^1.21.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-django = "^4.5.2"
coverage = "^6.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
